+++ crates/trading/src/post.rs
@@ -546,7 +546,7 @@ where
     )
     .await?;
 
-    let preview_from = tx.order_to_sign.receiver.clone();
+    let preview_from = tx.from.clone();
     let preview = OrderCreation::new(
         tx.order_to_sign.sell_token.clone(),
         tx.order_to_sign.buy_token.clone(),